- 3.2.4 beta 1:   - ALPHA!   - Some bugfixes and refactorings by Tobias.   - Started (work in progress) cleanup of includes.   - Many more refactorings by aschoelzhorn. Thanks again!     - Seperated wifi, temperature and other code to dedicated files.     - introduces rancilio-enums.h file and replaced some defines with enum classes.     - reorganized OTA (init and handle), OTA code still in "rancilio-pid.ino" because of timer stuff.     - separated declaration and implementation in some header/code files.
