+++ .github/workflows/platformio-build.yml
@@ -3,7 +3,21 @@ name: Test build
 on: [push, pull_request]
 
 jobs:
-    build:
+    nodemcuv2_ota:
+      runs-on: ubuntu-latest
+      steps:
+        - name: Checkout
+          uses: actions/checkout@v3
+        - name: Add config file
+          working-directory: ./rancilio-pid
+          run: cp userConfig.h.SAMPLE userConfig.h
+        - name: PlatformIO Run nodemcuv2_ota
+          uses: karniv00l/platformio-run-action@v1
+          with:
+            environments: "nodemcuv2_ota"
+            silent: false
+            verbose: true
+    esp32_ota_LIVE:
       runs-on: ubuntu-latest
       steps:
         - name: Checkout
@@ -17,10 +31,5 @@ jobs:
             environments: "esp32_ota_LIVE"
             silent: false
             verbose: true
-        - name: PlatformIO Run nodemcuv2_ota
-          uses: karniv00l/platformio-run-action@v1
-          with:
-            environments: "nodemcuv2_ota"
-            silent: false
-            verbose: true
+
