--- rancilio-pid/connection.h
@@ -1,13 +1,0 @@
-#ifndef _connection_H
-#define _connection_H
-
-extern bool forceOffline;
-
-void InitWifi(bool);
-void checkWifi(bool);
-void HandleOTA();
-void InitOTA();
-
-extern void disableBlynkTemporary();
-
-#endif // _connection_H
