+++ rancilio-pid/eeprom-pcpid.ino
@@ -4,6 +4,7 @@
  *****************************************************/
 
 #include "eeprom-pcpid.h"
+#include "rancilio-debug.h"
 #include "rancilio-pid.h"
 
 void sync_eeprom() { sync_eeprom(false, false); }
