+++ rancilio-pid/icon_winter.h
@@ -1,3 +1,7 @@
+/********************************************************
+ * Perfect Coffee PID
+ * https://github.com/medlor/bleeding-edge-ranciliopid
+ *****************************************************/
 #define icon_width 45
 #define icon_height 45
