+++ rancilio-pid/rancilio-debug.cpp
@@ -1,0 +1,9 @@
+/********************************************************
+ * Perfect Coffee PID
+ * https://github.com/medlor/bleeding-edge-ranciliopid
+ *****************************************************/
+#include "rancilio-debug.h"
+
+RemoteDebug Debug;
+
+
