+++ rancilio-pid/userConfig.h.SAMPLE
@@ -259,7 +259,6 @@
 #define MQTT_PASSWORD "mypass"
 #define MQTT_TOPIC_PREFIX "custom/Küche."  // topic will be "<MQTT_TOPIC_PREFIX><HOSTNAME>/<READING>" <HOSTNAME> is defined in //Wifi section
 #define MQTT_SERVER_IP "192.168.1.1"       // (MQTT_ENABLE=1) IP-Address of locally installed mqtt server
-#define MQTT_SERVER_PORT "1883"            // (MQTT_ENABLE=1)
-#define MQTT_MAX_PUBLISH_SIZE 120 
+#define MQTT_SERVER_PORT 1883              // (MQTT_ENABLE=1)
 
 #endif // _userConfig_H
